Import py27-dimarray-0.1.8 as wip/py-dimarray.
dimarray is a package to handle numpy arrays with labelled dimensions and axes
Inspired from pandas, it includes advanced alignment and reshaping features
and as well as missing-value (NaN) handling.

The main difference with pandas is that it is generalized to N dimensions, and
behaves more closely to a numpy array. The axes do not have fixed names
('index', 'columns', etc...) but are given a meaningful name by the user
(e.g. 'time', 'items', 'lon' ...). This is especially useful for high
dimensional problems such as sensitivity analyses.

A natural I/O format for such an array is netCDF, common in geophysics,
which relies on the netCDF4 package, and supports metadata.CVS: ----------------------------------------------------------------------