(this package is IMCOMPLETE, purely on the way) Import gnuradio-core-3.7.2.1 as wip/gnuradio-core.
GNU Radio is a collection of software that when combined with minimal hardware,
allows the construction of radios where the actual waveforms transmitted and
received are defined by software. What this means is that it turns the digital
modulation schemes used in today's high performance wireless devices into
software problems.

This module provides the main library containing the underlying runtime
system and most of the hardware independent signal processing blocks.