Import ii-1.7 as wip/ii.
ii is a minimalist FIFO and filesystem-based IRC client. It creates an irc
directory tree with server, channel and nick name directories.
In every directory a FIFO in file and a normal out file is created.

The in file is used to communicate with the servers and the out files
contain the server messages. For every channel and every nick name there
are related in and out files created. This allows IRC communication from
command line and adheres to the Unix philosophy.