Import py27-qnet-1.1.9 as wip/py-qnet.
The QNET package is a set of tools created and compiled to aid
in the design and analysis of photonic circuit models.
Our proposed Quantum Hardware Description Language (cf. QHDL_)
serves to describe a circuit topology and specification of a larger
entity in terms of parametrizable subcomponents.
This is strongly analogous to the specification of electric circuitry
using the structural description elements of VHDL or Verilog.
The physical systems that can be modeled within the framework include
quantum optical experiments that can be described as nodes with internal
degrees of freedom such as interacting quantum harmonic oscillators
and/or N-level quantum systems that,
in turn are coupled to a finite number of external bosonic quantum fields.