import of libqtlua for qt4 to pkgsrc-wip as qt4-libqtlua.
The QtLua library aims to make Qt4 applications scriptable using the
Lua scripting language. It is an alternative to the QtScript module.

QtLua does not generate or use generated binding code for Qt. Instead it
provides useful C++ wrapper classes to make C++ and lua objects
accessible from both lua and C++. It takes advantages of the Qt meta
object system to expose QObject members to lua scripts.

QtLua is not designed to develop an application in pure Lua but rather
to extend a Qt C++ application using Lua as scripting language.