(devel/SOPE to update 2.2.5 to 2.2.10, but Release info is not easy to find, so the wip directory)
Import SOPE-2.2.10 as wip/SOPE.

SOPE stands for "SKYRiX Object Publishing Environment" and is a
framework for mapping dynamic objects into the HTTP infrastructure,
that is, the web. A lot of ideas in SOPE are actually inspired by
ZOPE (the Z Object Publishing Environment), but the implementation
and details are quite different.

This version of SOPE is maintained by SOGo team.