Import mate-indicator-applet-1.8.0 as wip/mate-indicator-applet.
A small applet to display information from various applications consistently in
the panel.

The indicator applet exposes Ayatana Indicators in the MATE Panel. Ayatana
Indicators are an initiative by Canonical to provide crisp and clean system and
application status indication. They take the form of an icon and associated
menu, displayed (usually) in the desktop panel. Existing indicators include the
Message Menu, Battery Menu and Sound menu.

MATE Indicator Applet is a fork of Indicator Applet for GNOME.