Update 2.0.3 to 2.1RC3  (it is said beta-3 at following text, but we call it as RC) (Following text is note from 2.1 beta-1 to 2.1-beta-3) ------------------------------ I'm happy to announce availability of a test release of TQSL, which is the third beta test release for TQSL 2.1.
The two major goals for TQSL 2.1 are to make TQSL accessible and to make it localizable.

Changes since beta-2:

1. Several cases where errors could either be ignored or lost have been corrected. For
example, trying to create a callsign certificate request when the user does not have
permission to save the request file would result in "NO ERROR" being reported. Trying to
import a .TQ6 file when the user did not have permission to store the callsign certificate
would cause TQSL to crash.

2. The certificate upload messages from LoTW are now cleaned up prior to display. When an
error occurs during certificate processing for an uploaded certificate request, TQSL now
displays that error more prominently and deletes the pending callsign certificate request.

3. Recovery from errors in the Duplicates database have been made more robust. The common
case where TQSL would report "Invalid Argument" (trying to use a database from a newer
version of TQSL on an old version of TQSL) is now handled.

(These are also available from the Sourceforge repository.)

As always, feedback on any suggestions for improvements or reports of any issues is
appreciated. Send directly to me, or on the list

[The TQSL 2.1 beta-2 release notes follow.]

For accessibility, several changes in the TQSL user interface were made to help the user
interface to interact better with screen readers and other accessibility interfaces. While
many of these changes aren't apparent to the typical user, there are some minor changes in
work flow that you may notice. However, it is much more usable with assistive technologies.

For localization, much of the TQSL GUI and messages can now be translated into local
languages. For this beta, there are machine translations included that came from Google
Translate. They're probably of very poor quality, but serve as a proof of concept. Just use
the "Language" selection on the "File" pulldown menu. Note that you may need extra fonts
installed to use some languages.

Other major changes:

1. If a user has only one station location, then there's no prompt to ask them to select what
location to use when signing a log.

2. Station locations can be set up with callsign "[None]". When a user uses that station
location to sign a log, they're prompted for the appropriate callsign. This allows people who
have had multiple callsigns for a given QTH to enter the data for the location once and
re-use it for a specific callsign.

3. TQSL now recognizes the ADIF <SUBMODE> tag and attempts to map that to modes recognized by
LoTW. Fully supporting <SUBMODE> will require changes in the LoTW configuration file.

4. TQSL now detects when another instance is busy signing a log (waiting on a user to respond
to a prompt, for example) and a command line logger is requesting TQSL to sign a different
log. The command line invocation of TQSL no longer "hangs" the command line signing operation
and instead returns an error.

5. TQSL now properly handles non-ASCII characters in folder names. Previous versions of TQSL
relied upon the DOS compatibility "short" file names, which are not enabled by default on
Windows 8.

There's also startup performance improvements, particularly for stations with large numbers
of callsign certificates. Other minor updates include allowing callsign certificates to be
requested with dates one year in the future (for special event stations, for example), and
verifying that the end date for valid QSOs for a callsign certificate is after the start
date.