(Candidate update devel/p5-SDL from 2.1.2 to 2.544) Import p5-SDL-2.544 as wip/p5-SDL.
SDL_perl is a package of Perl modules that provide both functional and
object oriented interfaces to the Simple DirectMedia Layer for Perl 5.

This package takes some liberties with the SDL API, and attempts to adhere
to the spirit of both the SDL and Perl. This document describes the
low-level functional SDL Perl API. For the object oriented programming
interface please see the documentation provided on a per-class basis.CVS: ----------------------------------------------------------------------