Import phantomjs-2.0.0 as wip/phantomjs2.
PhantomJS is a headless WebKit with JavaScript API. It has fast and native
support for various web standards: DOM handling, CSS selector, JSON, Canvas,
SVG. PhantomJS is an optimal solution for fast headless testing, site
scraping, pages capture, network monitoring and many other use cases.

This is the current stable version. Please note that this still does not work
and build on - at least - NetBSD/amd64 7.99.5, pkgsrc-current. Discussed in a
private email with Filip Hajny. For more information regarding the issues please
give a look to the TODO and the various XXX and FIXME in the Makefile.