Import syncthing-inotify-0.5 as wip/syncthing-inotify.
Syncthing (core) uses a rescan interval to detect changes in folders. This
application (syncthing-inotify) uses OS primitives to detect changes as soon
as they happen. Therefore, if you save a file, syncthing-inotify will know
about it and pass this information to Syncthing such that near real-time
synchronisation can be achieved.