(Based on PR pkg/49705 http://gnats.netbsd.org/49705 by tsutui, update candidate for adobe-flash-plugin)
Import adobe-flash-plugin-11.2.202.451 as wip/adobe-flash-plugin11.

Netscape compatible plugin for Adobe Flash player.  This package
contains a plugin that enables web browsers to render the Flash
format.

The nspluginwrapper option (enabled by default on non Linux platforms)
permits the use of the the Linux x86 flash player in Firefox browsers
on other x86 systems and x86_64 systems.

For NetBSD this package requires emul.linux.kern.osrelease = 2.6.18
or later, which means it will not work on NetBSD 5.x or older
systems.