Import py27-chemlab-0.3 as wip/py-chemlab.
chemlab is a python library and a set of utilities built to ease the
life of the computational chemist. It takes inspiration from other
python scientific library such as numpy scipy and matplotlib, and aims
to bring a consistent and simple API by following the python
guidelines.

Computational and theoretical chemistry is a huge
field, and providing a program that encompasses all aspects of it is an
impossible task. The spirit of chemlab is to provide a common ground
from where you can build specific programs. For this reason it
includes an easily extendable molecular viewer and flexible data
structures field-independent