Import memo-1.6 as wip/memo.
Memo is a note-taking and to-do software written in C. Memo runs on command
line.

With Memo, it's possible to concentrate on the tasks in your to-do list
instead of spending time managing those tasks. After a while taking notes
and keeping a to-do list is part of your daily life instead of some nasty
thing you must remember to do.

Memo is Free Software under the GPL licence and is maintained by Niko
Rosvall.